Add CbusBasicNode etc to extend upon
